requirements.txt
transformers>=4.33.0
accelerate>=0.22.0
huggingface_hub>=0.16.0
diffusers>=0.24.0
tqdm>=4.66.0
safetensors>=0.3.2
pillow>=9.5.0
requests>=2.28.0
